show cursos,validation,edit likprof

+++ database/migrations/2022_11_15_201124_create_curso_user_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('curso_user', function (Blueprint $table) {
             $table->foreignId('curso_id')->constrained()->onDelete('cascade');;
             $table->foreignId('user_id')->constrained()->onDelete('cascade');;
-            $table->integer('nota')->nullable();
+            $table->float('nota')->nullable();
             $table->timestamps();
         });
     }
